# frontend/openapi/wst-openapi.yaml
openapi: 3.1.0
info:
  title: Workshop Management & Student Practical Training (WST) API
  version: 1.0.0
  summary: Contract-first REST API for the WST academic MVP.
  description: |
    Source of truth for the PostgreSQL schema, Node.js backend, automated tests, React client and documentation.
    Requirement prefix: WST. Every operation carries `x-requirement-id` and `x-priority`.

    ## Conventions
    - Base path `/api/v1`. JSON bodies (multipart only for attachment upload). Bearer JWT. Operations with `security: []` are public.
    - Identifiers are UUIDs. Human-readable numbers (`jobNumber`, `poNumber`, `invoiceNumber`, `certificateNumber`) are server-generated, unique, immutable, and never localized or reordered by the backend.
    - Timestamps are RFC 3339 UTC. Dates are `YYYY-MM-DD`.
    - Money is `{amount, currency}`: `amount` is a decimal string (never a float) and `currency` an ISO-4217 code. One system currency is configured in finance settings; documents in another currency are rejected with 422. Part quantities are whole numbers. Rates and metric values are decimal strings.
    - Optional properties without a value are omitted, never null. `readOnly` properties are server-controlled. Request schemas forbid unknown properties, so clients cannot submit ids, audit fields, stock balances, invoice totals, or permissions.
...

    ## Collections
...

    ## Authorization
...

    ## Concurrency and idempotency
...

    ## Localization
...

    ## Errors and rate limiting
...

...

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /technicians:
    get:
      tags: [Workshop Jobs]
      operationId: listTechnicians
      summary: Directory of assignable technicians
      description: "Requires `jobs.assign`. Users holding the TECHNICIAN role within the caller's scopes. Sortable: displayName."
      x-requirement-id: [WST-FR-04]
      x-priority: Must
      x-permissions: [jobs.assign]
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/Sort'
        - $ref: '#/components/parameters/Query'
      responses:
        '200': { description: Technicians, content: { application/json: { schema: { $ref: '#/components/schemas/UserRefPage' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /job-cards:
    get:
      tags: [Workshop Jobs]
      operationId: listJobCards
      summary: List job cards
      description: "Requires `jobs.read` or `jobs.read.assigned`. Technicians only see jobs assigned to them; other rows return as absent. Sortable: jobNumber, createdAt, expectedCompletionAt, priority (default -createdAt)."
      x-requirement-id: [WST-FR-04, WST-FR-01]
      x-priority: Must
      x-permissions: [jobs.read, jobs.read.assigned]
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/Sort'
        - $ref: '#/components/parameters/Query'
        - $ref: '#/components/parameters/From'
        - $ref: '#/components/parameters/To'
        - { name: stage, in: query, schema: { $ref: '#/components/schemas/JobStage' } }
        - { name: priority, in: query, schema: { $ref: '#/components/schemas/JobPriority' } }
        - { name: jobNumber, in: query, schema: { type: string, maxLength: 20 } }
        - { name: technicianId, in: query, schema: { $ref: '#/components/schemas/Uuid' } }
        - { name: bayId, in: query, schema: { $ref: '#/components/schemas/Uuid' } }
        - { name: vehicleId, in: query, schema: { $ref: '#/components/schemas/Uuid' } }
        - { name: customerId, in: query, schema: { $ref: '#/components/schemas/Uuid' } }
      responses:
        '200': { description: Job cards, content: { application/json: { schema: { $ref: '#/components/schemas/JobCardPage' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Workshop Jobs]
      operationId: createJobCard
      summary: Open a job card
      description: "Requires `jobs.create`. Required fields are enforced (422). The server generates a unique human-readable jobNumber, sets stage RECEIVED, derives the customer from the vehicle, rejects archived vehicles and intake mileage below the recorded mileage, links the caller's uploaded attachments, and writes the first stage event and an audit event."
      x-requirement-id: [WST-FR-04, WST-FR-05]
      x-priority: Must
      x-permissions: [jobs.create]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/JobCardCreateRequest' }
            examples:
              basic: { $ref: '#/components/examples/JobCardCreateExample' }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/JobCard' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /job-cards/{jobId}:
    parameters:
      - $ref: '#/components/parameters/JobId'
    get:
      tags: [Workshop Jobs]
      operationId: getJobCard
      summary: Get a job card
      description: "Requires `jobs.read` or `jobs.read.assigned` (assigned jobs only)."
      x-requirement-id: [WST-FR-04]
      x-priority: Must
      x-permissions: [jobs.read, jobs.read.assigned]
      responses:
        '200': { description: Job card, content: { application/json: { schema: { $ref: '#/components/schemas/JobCard' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [Workshop Jobs]
      operationId: updateJobCard
      summary: Update permitted job fields
      description: "Requires `jobs.update`. Cannot change stage, jobNumber, vehicle or customer, and is rejected for DELIVERED jobs (409 JOB_STAGE_NOT_ALLOWED). Stage changes happen only through the transitions resource."
      x-requirement-id: [WST-FR-04]
      x-priority: Must
      x-permissions: [jobs.update]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/JobCardUpdateRequest' } } } }
      responses:
        '200': { description: Updated, content: { application/json: { schema: { $ref: '#/components/schemas/JobCard' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /job-cards/{jobId}/assignment:
    parameters:
      - $ref: '#/components/parameters/JobId'
    put:
      tags: [Workshop Jobs]
      operationId: assignJobCard
      summary: Assign bay, technician and time window
      description: "Requires `jobs.assign`. Validates the shared bay calendar server-side: bay availability, overlap with other active jobs and with published training sessions, and technician double-booking as a mentor. Conflicts return 409 SCHEDULE_CONFLICT with an explainable conflict list; job assignments have no override. Not allowed for DELIVERED jobs."
      x-requirement-id: [WST-FR-04, WST-FR-10]
      x-priority: Must
      x-permissions: [jobs.assign]
      parameters:
        - $ref: '#/components/parameters/AcceptLanguage'
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/JobAssignmentRequest' } } } }
      responses:
        '200': { description: Assigned, content: { application/json: { schema: { $ref: '#/components/schemas/JobCard' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/ScheduleConflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /job-cards/{jobId}/transitions:
    parameters:
      - $ref: '#/components/parameters/JobId'
    post:
      tags: [Workshop Jobs]
      operationId: transitionJobCard
      summary: Move a job to the next lifecycle stage
      description: |
        Controlled transition; there is no PATCH of stage. Allowed transitions only: RECEIVED to IN_PROGRESS, IN_PROGRESS to QUALITY_CHECK, QUALITY_CHECK to IN_PROGRESS (rework), QUALITY_CHECK to READY, READY to DELIVERED. Anything else returns 409 INVALID_STATE_TRANSITION. If the current stage differs from expectedFromStage the result is 409 VERSION_CONFLICT.
        Permission by target: IN_PROGRESS from RECEIVED needs `jobs.transition.start`; QUALITY_CHECK needs `jobs.transition.submit-qc`; IN_PROGRESS from QUALITY_CHECK needs `quality.perform`; READY needs `jobs.transition.ready`; DELIVERED needs `jobs.transition.deliver`. Technicians act only on assigned jobs.
        Guards: start requires bay and technician (JOB_ASSIGNMENT_REQUIRED) and an APPROVED INITIAL_WORK approval (CUSTOMER_APPROVAL_REQUIRED). Submitting to quality check requires every non-cancelled work item to be DONE (CHECKLIST_INCOMPLETE). Rework requires the latest quality check to be FAILED and a reason. READY requires the latest quality check to be PASSED (QUALITY_CHECK_REQUIRED) and atomically generates the DRAFT invoice from logged labor and issued parts. DELIVERED requires a non-void invoice (INVOICE_REQUIRED).
        Each success appends an immutable stage event (responsible user, timestamp, previous stage, new stage) and an audit event.
      x-requirement-id: [WST-FR-05, WST-FR-01]
      x-priority: Must
      x-permissions: [jobs.transition.start, jobs.transition.submit-qc, quality.perform, jobs.transition.ready, jobs.transition.deliver]
      parameters:
        - $ref: '#/components/parameters/AcceptLanguage'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/JobTransitionRequest' }
            examples:
              start: { $ref: '#/components/examples/JobTransitionExample' }
      responses:
        '200': { description: Transitioned, content: { application/json: { schema: { $ref: '#/components/schemas/JobCard' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /job-cards/{jobId}/stage-history:
    parameters:
      - $ref: '#/components/parameters/JobId'
    get:
      tags: [Workshop Jobs]
      operationId: listJobStageHistory
      summary: Complete auditable stage timeline
      description: "Requires `jobs.read` or `jobs.read.assigned`. Immutable events, oldest first by default. Sortable: transitionedAt."
      x-requirement-id: [WST-FR-05]
      x-priority: Must
      x-permissions: [jobs.read, jobs.read.assigned]
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/Sort'
      responses:
        '200': { description: Timeline, content: { application/json: { schema: { $ref: '#/components/schemas/JobStageEventPage' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /job-cards/{jobId}/work-items:
    parameters:
      - $ref: '#/components/parameters/JobId'
    get:
      tags: [Workshop Jobs]
      operationId: listWorkItems
      summary: List the work checklist
      description: "Requires `jobs.read` or `jobs.read.assigned`. Sortable: createdAt."
      x-requirement-id: [WST-FR-04]
      x-priority: Must
      x-permissions: [jobs.read, jobs.read.assigned]
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/Sort'
        - { name: status, in: query, schema: { type: string, enum: [PENDING, DONE, CANCELLED] } }
      responses:
        '200': { description: Work items, content: { application/json: { schema: { $ref: '#/components/schemas/WorkItemPage' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      tags: [Workshop Jobs]
      operationId: createWorkItem
      summary: Add a checklist item
      description: "Requires `jobs.update`. Allowed in RECEIVED or IN_PROGRESS. isAdditionalWork items cannot receive labor or parts until an APPROVED ADDITIONAL_WORK approval references them."
      x-requirement-id: [WST-FR-04]
      x-priority: Must
      x-permissions: [jobs.update]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/WorkItemCreateRequest' } } } }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/WorkItem' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /job-cards/{jobId}/work-items/{workItemId}:
    parameters:
      - $ref: '#/components/parameters/JobId'
      - $ref: '#/components/parameters/WorkItemId'
    patch:
      tags: [Workshop Jobs]
      operationId: updateWorkItem
      summary: Update checklist item status or text
      description: "Requires `labor.write` (status changes by the assigned technician) or `jobs.update` (text and cancellation). Allowed in IN_PROGRESS for status changes."
      x-requirement-id: [WST-FR-04]
      x-priority: Must
      x-permissions: [labor.write, jobs.update]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/WorkItemUpdateRequest' } } } }
      responses:
        '200': { description: Updated, content: { application/json: { schema: { $ref: '#/components/schemas/WorkItem' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /job-cards/{jobId}/quality-checks:
    parameters:
      - $ref: '#/components/parameters/JobId'
    get:
      tags: [Workshop Jobs]
      operationId: listQualityChecks
      summary: List quality checks of a job
      description: "Requires `jobs.read`, `jobs.read.assigned` or `quality.perform`. Sortable: performedAt."
      x-requirement-id: [WST-FR-05]
      x-priority: Must
      x-permissions: [jobs.read, jobs.read.assigned, quality.perform]
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/Sort'
      responses:
        '200': { description: Quality checks, content: { application/json: { schema: { $ref: '#/components/schemas/QualityCheckPage' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      tags: [Workshop Jobs]
      operationId: createQualityCheck
      summary: Record a quality-check result with evidence
      description: "Requires `quality.perform`, a permission distinct from job execution. Job must be in QUALITY_CHECK (409 JOB_STAGE_NOT_ALLOWED). A FAILED result requires notes. Recording the result does not move the stage; use the transitions resource to progress (READY) or reject (IN_PROGRESS). Audited."
      x-requirement-id: [WST-FR-05, WST-FR-04]
      x-priority: Must
      x-permissions: [quality.perform]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/QualityCheckCreateRequest' } } } }
      responses:
        '201': { description: Recorded, content: { application/json: { schema: { $ref: '#/components/schemas/QualityCheck' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Short-lived access token issued by login or refresh. The token identifies the user only; roles, permissions and scopes are resolved server-side on each request."
  parameters:
    Page: { name: page, in: query, description: "1-based page number.", schema: { type: integer, minimum: 1, default: 1 } }
    PageSize: { name: pageSize, in: query, description: "Items per page, maximum 100.", schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
    Sort: { name: sort, in: query, description: "Comma-separated fields; prefix - for descending. Only fields named by the operation are accepted.", schema: { type: string, maxLength: 100, pattern: '^-?[A-Za-z][A-Za-z0-9]*(,-?[A-Za-z][A-Za-z0-9]*)*$' } }
    From: { name: from, in: query, description: "Inclusive lower bound (RFC 3339 UTC).", schema: { $ref: '#/components/schemas/Timestamp' } }
    To: { name: to, in: query, description: "Exclusive upper bound (RFC 3339 UTC).", schema: { $ref: '#/components/schemas/Timestamp' } }
    Query: { name: q, in: query, description: "Case-insensitive text search.", schema: { type: string, minLength: 1, maxLength: 100 } }
    AcceptLanguage: { name: Accept-Language, in: header, description: "en or ar. Selects the language of messages and labels; unsupported values fall back to en. Never alters identifiers or amounts.", schema: { $ref: '#/components/schemas/Locale' } }
...
    JobId: { name: jobId, in: path, required: true, schema: { $ref: '#/components/schemas/Uuid' } }
    WorkItemId: { name: workItemId, in: path, required: true, schema: { $ref: '#/components/schemas/Uuid' } }
...
  responses:
    BadRequest:
      description: "Malformed request, unknown filter or sort field, or invalid query parameter."
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Unauthorized:
      description: "Missing, invalid or expired credentials."
      headers: { WWW-Authenticate: { schema: { type: string } } }
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Forbidden:
      description: "Authenticated but lacking the required permission. Rows outside the caller's scope are reported as 404 instead."
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    NotFound:
      description: "Resource does not exist or is outside the caller's row scope."
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Conflict:
      description: "Invalid state transition, missing customer approval, insufficient stock, concurrency conflict, separation-of-duties violation or duplicate business operation. See Error.code."
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    ScheduleConflict:
      description: "Shared bay or mentor scheduling conflict, or another conflict per Error.code. Scheduling conflicts include an explainable list."
      content: { application/json: { schema: { $ref: '#/components/schemas/ScheduleConflictError' }, examples: { conflict: { $ref: '#/components/examples/ScheduleConflictExample' } } } }
    ValidationError:
      description: "Semantic validation failed; details identify each offending field."
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
  examples:
    JobCardCreateExample:
      summary: Open a periodic-service job
      value: { vehicleId: 3f2a6d1e-8b7c-4c1d-9a55-0d6f1a2b3c4d, complaint: "Engine warning light and noise when braking", serviceType: REPAIR, priority: NORMAL, mileageAtIntake: 84210, expectedCompletionAt: '2026-10-02T15:00:00Z', workItems: [ { description: "Diagnose warning light" }, { description: "Inspect brake pads" } ] }
    JobTransitionExample:
      summary: Start work
      value: { toStage: IN_PROGRESS, expectedFromStage: RECEIVED }
...
  schemas:
    Uuid: { type: string, format: uuid }
    Timestamp: { type: string, format: date-time, description: "RFC 3339 UTC timestamp." }
...
    Locale: { type: string, enum: [en, ar] }
...
    RecordMeta:
      type: object
      required: [id, createdAt, updatedAt, createdBy, updatedBy]
      properties:
        id: { $ref: '#/components/schemas/Uuid', readOnly: true }
        createdAt: { $ref: '#/components/schemas/Timestamp', readOnly: true }
        updatedAt: { $ref: '#/components/schemas/Timestamp', readOnly: true }
        createdBy: { $ref: '#/components/schemas/Uuid', readOnly: true }
        updatedBy: { $ref: '#/components/schemas/Uuid', readOnly: true }
    PageInfo:
      type: object
      required: [page, pageSize, totalItems, totalPages]
      properties: { page: { type: integer, minimum: 1 }, pageSize: { type: integer, minimum: 1, maximum: 100 }, totalItems: { type: integer, minimum: 0 }, totalPages: { type: integer, minimum: 0 } }
...
    UserRef:
      type: object
      required: [id, displayName]
      properties: { id: { $ref: '#/components/schemas/Uuid' }, displayName: { type: string } }
...
    Error:
      type: object
      required: [code, message, requestId]
      properties:
        code: { $ref: '#/components/schemas/ErrorCode' }
        message: { type: string, description: "Human-readable message in the negotiated language." }
        details: { type: array, items: { $ref: '#/components/schemas/ErrorDetail' } }
        requestId: { type: string, description: "Correlation id also present in logs and audit events." }
    ScheduleConflictError:
      allOf:
        - $ref: '#/components/schemas/Error'
        - type: object
          required: [conflicts]
          properties: { conflicts: { type: array, items: { $ref: '#/components/schemas/ScheduleConflict' } } }
...
    JobNumber: { type: string, pattern: '^JC-[0-9]{4}-[0-9]{6}$', description: "Server-generated, unique, human-readable. Always rendered LTR by clients." }
    JobStage: { type: string, enum: [RECEIVED, IN_PROGRESS, QUALITY_CHECK, READY, DELIVERED] }
    JobPriority: { type: string, enum: [LOW, NORMAL, HIGH, URGENT] }
    ServiceType: { type: string, enum: [MAINTENANCE, REPAIR, DIAGNOSTIC, INSPECTION, OTHER] }
    JobApprovalSummary:
      type: object
      required: [billableWorkAllowed, approvedScopes, pendingApprovalCount]
      properties: { billableWorkAllowed: { type: boolean, description: "True only when an APPROVED INITIAL_WORK approval exists." }, approvedScopes: { type: array, items: { $ref: '#/components/schemas/ApprovalScope' }, uniqueItems: true }, pendingApprovalCount: { type: integer, minimum: 0 } }
    JobCard:
      allOf:
        - $ref: '#/components/schemas/RecordMeta'
        - type: object
          required: [jobNumber, customerId, vehicleId, organizationScopeId, complaint, serviceType, priority, mileageAtIntake, stage, expectedCompletionAt, approvalSummary, version]
          properties:
            jobNumber: { $ref: '#/components/schemas/JobNumber' }
            customerId: { $ref: '#/components/schemas/Uuid' }
            customerDisplayName: { type: string }
            vehicleId: { $ref: '#/components/schemas/Uuid' }
            vehiclePlate: { type: string }
            organizationScopeId: { $ref: '#/components/schemas/Uuid' }
            complaint: { type: string, minLength: 3, maxLength: 2000 }
            serviceType: { $ref: '#/components/schemas/ServiceType' }
            priority: { $ref: '#/components/schemas/JobPriority' }
            mileageAtIntake: { type: integer, minimum: 0 }
            stage: { $ref: '#/components/schemas/JobStage' }
            stageChangedAt: { $ref: '#/components/schemas/Timestamp' }
            scheduledStartAt: { $ref: '#/components/schemas/Timestamp' }
            expectedCompletionAt: { $ref: '#/components/schemas/Timestamp' }
            bayId: { $ref: '#/components/schemas/Uuid' }
            technicianId: { $ref: '#/components/schemas/Uuid' }
            approvalSummary: { $ref: '#/components/schemas/JobApprovalSummary' }
            deliveredAt: { $ref: '#/components/schemas/Timestamp' }
            version: { type: integer, minimum: 1 }
    JobCardCreateRequest:
      type: object
      required: [vehicleId, complaint, serviceType, priority, mileageAtIntake, expectedCompletionAt]
      additionalProperties: false
      properties:
        vehicleId: { $ref: '#/components/schemas/Uuid' }
        complaint: { type: string, minLength: 3, maxLength: 2000 }
        serviceType: { $ref: '#/components/schemas/ServiceType' }
        priority: { $ref: '#/components/schemas/JobPriority' }
        mileageAtIntake: { type: integer, minimum: 0 }
        expectedCompletionAt: { $ref: '#/components/schemas/Timestamp' }
        workItems: { type: array, maxItems: 50, items: { $ref: '#/components/schemas/WorkItemCreateRequest' } }
        attachmentIds: { type: array, maxItems: 10, items: { $ref: '#/components/schemas/Uuid' }, uniqueItems: true }
    JobCardUpdateRequest:
      type: object
      required: [version]
      additionalProperties: false
      properties: { version: { type: integer, minimum: 1 }, complaint: { type: string, minLength: 3, maxLength: 2000 }, priority: { $ref: '#/components/schemas/JobPriority' }, serviceType: { $ref: '#/components/schemas/ServiceType' }, expectedCompletionAt: { $ref: '#/components/schemas/Timestamp' } }
    JobAssignmentRequest:
      type: object
      required: [version, bayId, technicianId, scheduledStartAt, expectedCompletionAt]
      additionalProperties: false
      properties: { version: { type: integer, minimum: 1 }, bayId: { $ref: '#/components/schemas/Uuid' }, technicianId: { $ref: '#/components/schemas/Uuid' }, scheduledStartAt: { $ref: '#/components/schemas/Timestamp' }, expectedCompletionAt: { $ref: '#/components/schemas/Timestamp' } }
    JobTransitionRequest:
      type: object
      required: [toStage, expectedFromStage]
      additionalProperties: false
      properties: { toStage: { $ref: '#/components/schemas/JobStage' }, expectedFromStage: { $ref: '#/components/schemas/JobStage' }, reason: { type: string, minLength: 3, maxLength: 500, description: "Required when returning from QUALITY_CHECK to IN_PROGRESS." } }
...
    WorkItem:
      allOf:
        - $ref: '#/components/schemas/RecordMeta'
        - type: object
          required: [jobId, description, status, isAdditionalWork]
          properties: { jobId: { $ref: '#/components/schemas/Uuid' }, description: { type: string, maxLength: 300 }, status: { type: string, enum: [PENDING, DONE, CANCELLED] }, isAdditionalWork: { type: boolean }, approvalId: { $ref: '#/components/schemas/Uuid' }, completedAt: { $ref: '#/components/schemas/Timestamp' } }
    WorkItemCreateRequest:
      type: object
      required: [description]
      additionalProperties: false
      properties: { description: { type: string, minLength: 1, maxLength: 300 }, isAdditionalWork: { type: boolean, default: false } }
    WorkItemUpdateRequest:
      type: object
      minProperties: 1
      additionalProperties: false
      properties: { description: { type: string, minLength: 1, maxLength: 300 }, status: { type: string, enum: [PENDING, DONE, CANCELLED] } }
    QualityCheck:
      type: object
      required: [id, jobId, result, performedBy, performedAt, evidenceAttachmentIds]
      properties: { id: { $ref: '#/components/schemas/Uuid' }, jobId: { $ref: '#/components/schemas/Uuid' }, result: { type: string, enum: [PASSED, FAILED] }, notes: { type: string, maxLength: 2000 }, evidenceAttachmentIds: { type: array, items: { $ref: '#/components/schemas/Uuid' } }, performedBy: { $ref: '#/components/schemas/Uuid' }, performedAt: { $ref: '#/components/schemas/Timestamp' } }
    QualityCheckCreateRequest:
      type: object
      required: [result]
      additionalProperties: false
      properties: { result: { type: string, enum: [PASSED, FAILED] }, notes: { type: string, maxLength: 2000 }, evidenceAttachmentIds: { type: array, maxItems: 10, items: { $ref: '#/components/schemas/Uuid' }, uniqueItems: true } }
    ApprovalScope: { type: string, enum: [INITIAL_WORK, ADDITIONAL_WORK, SUBLET] }
...
    UserRefPage: { type: object, required: [items, page], properties: { items: { type: array, items: { $ref: '#/components/schemas/UserRef' } }, page: { $ref: '#/components/schemas/PageInfo' } } }
...
    JobCardPage: { type: object, required: [items, page], properties: { items: { type: array, items: { $ref: '#/components/schemas/JobCard' } }, page: { $ref: '#/components/schemas/PageInfo' } } }
    JobStageEventPage: { type: object, required: [items, page], properties: { items: { type: array, items: { $ref: '#/components/schemas/JobStageEvent' } }, page: { $ref: '#/components/schemas/PageInfo' } } }
    WorkItemPage: { type: object, required: [items, page], properties: { items: { type: array, items: { $ref: '#/components/schemas/WorkItem' } }, page: { $ref: '#/components/schemas/PageInfo' } } }
    QualityCheckPage: { type: object, required: [items, page], properties: { items: { type: array, items: { $ref: '#/components/schemas/QualityCheck' } }, page: { $ref: '#/components/schemas/PageInfo' } } }
...